⭐ azure: wire IP-prefix/allocation/tap/private-endpoint refs into existing resources
Depth expansion + a file rename.

Cross-references (all live-verified against a real subscription):
- subnet.privateEndpoints() and subnet.ipAllocations(): the private
  endpoints deployed in a subnet and the IP allocations reserving its
  space.
- ipAddress.publicIpPrefix(): the public IP prefix an address was
  allocated from.
- interface.ipConfiguration.virtualNetworkTaps(): the taps mirroring an
  IP configuration's traffic.

Added inits for ipAllocation, publicIpPrefix, and virtualNetworkTap (and
extracted their builders) so the refs resolve by ARM ID; privateEndpoint
already had one.

Also renamed the meaningless network_tier3.go (an internal
prioritization label) into two domain files that match the provider's
by-domain convention: network_virtual_appliances.go (virtual appliance,
virtual router, virtual network tap) and network_ip_prefixes.go (public
IP prefix, custom IP prefix, IP allocation).
